Collate common helper functions into Julia module! v0.1.0
Over the years that I've been doing AoC, there have been a few
problems that use similar code, which I ended up copy and pasting.  I
figured it would be good to start collating these into a common
directory that I can import from.  The first of these functions are
related to multidimensional indexing, offsets, and such.  This commit
creates this module, and modifies some historical AoC solutions to use
this instead of copying and pasting.